// lisa-frontend/src/api/types.ts
export type UserRole = "AGRONOMIST" | "SALES_MANAGER" | "FARM_MANAGER"
// ...
export interface User {
  id: string
  fullName: string
  username: string
  role: UserRole
  createdAt?: string
  updatedAt?: string
}